columna facturacion añadida

+++ src/Notizap.Application/DTOs/Analisis/VentaDiaria/EvolucionSucursalResumenDto.cs
@@ -1,6 +1,6 @@
 public class EvolucionSucursalResumenDto
 {
     public string Sucursal { get; set; } = "";
-    public List<int> Serie { get; set; } = new();
+    public List<double> Serie { get; set; } = new();
     public string Color { get; set; } = "";
 }
